added demo files in gui/demo added channel selector ms work now -> problems with label/overlay update fixed (they tried to use the same channel as in the data)
git-svn-id: https://gorgonzola.iwr.uni-heidelberg.de/svn/ILASTIK/trunk/src@315 40e545cb-b269-0410-8983-4113551e4d47

+++ ilastikMain.py
@@ -211,7 +211,7 @@ class MainWindow(QtGui.QMainWindow):
                 if item.prediction is None:
                    item.prediction = numpy.zeros(imageItem.dataVol.data.shape[0:-1],'uint8')
                 if imageIndex == self.activeImage:
-                    color = QtGui.QColor.fromRgb(item.color)
+                    color = QtGui.QColor.fromRgb(long(item.color))
                     self.labelWidget.addOverlay(True, item.prediction, item.name, color, 0.4)
             
             if imageItem.dataVol.uncertainty is None:
